More Qt6 Porting - almost done
- ilandc is now "ported" too (easy)
- got rid of the compat-library by removing/changing instances of QRegExp to QRegularExpression.
- put an updated copy of Qt Highlighter example into jstextarea.cpp

Some tests comparing a simulation with qt5 and qt6 showed the following:
- Qt6 seems to need more RAM (for BGD landscape 11GB compared to 10GB)
- Qt6 runs faster:
-- javascript (devstageoutput): 30% faster in qt6 (9m 28.707s vs 12m 28.1815s), 
-- total: qt5: 27m 7.92626s, qt6: 23m 31.0944s (-> in addition to diff in devstageout, qt5 one minute faster)

Qt6 and Qt5 yielded numerically identical values for BGD landscape.
For a BITE test landscape not -> traced down to JS random number generation, therefore added Globals.random() function.

+++ src/core/svdstate.h
@@ -38,9 +38,9 @@ inline bool operator==(const SVDState &s1, const SVDState &s2)
     return equal;
 }
 
-inline uint qHash(const SVDState &key, uint seed)
+inline size_t qHash(const SVDState &key, size_t seed)
 {
-    uint hash_value = qHash(key.composition, seed) ^ qHash(key.structure, seed) ^ qHash(key.function, seed);
+    size_t hash_value = qHash(key.composition, seed) ^ qHash(key.structure, seed) ^ qHash(key.function, seed);
     return hash_value;
 }
